Named Colors : pushed into separate Util, created unit tests
